+++ src/gui/CMakeLists.txt
@@ -23,8 +23,9 @@ set(client_UI
     protocolwidget.ui
     settingsdialog.ui
     sharedialog.ui
-    shareusergroupdialog.ui
-    sharedialogshare.ui
+    sharelinkwidget.ui
+    shareusergroupwidget.ui
+    sharewidget.ui
     sslerrordialog.ui
     owncloudsetuppage.ui
     addcertificatedialog.ui
@@ -63,7 +64,8 @@ set(client_SRCS
     settingsdialog.cpp
     share.cpp
     sharedialog.cpp
-    shareusergroupdialog.cpp
+    sharelinkwidget.cpp
+    shareusergroupwidget.cpp
     sharee.cpp
     socketapi.cpp
     sslbutton.cpp
