+++ src/gui/sharedialog.h
@@ -1,6 +1,5 @@
 /*
  * Copyright (C) by Roeland Jago Douma <roeland@famdouma.nl>
- * Copyright (C) 2015 by Klaas Freitag <freitag@owncloud.com>
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -16,11 +15,9 @@
 #define SHAREDIALOG_H
 
 #include "accountfwd.h"
-#include "QProgressIndicator.h"
+#include <QString>
 #include <QDialog>
-#include <QVariantMap>
-#include <QSharedPointer>
-#include <QList>
+#include <QWidget>
 
 namespace OCC {
 
@@ -28,82 +25,37 @@ namespace Ui {
 class ShareDialog;
 }
 
-class AbstractCredentials;
-class QuotaInfo;
-class SyncResult;
-class LinkShare;
-class Share;
-class ShareManager;
+class ShareLinkWidget;
+class ShareUserGroupWidget;
 
-/**
- * @brief The ShareDialog class
- * @ingroup gui
- */
 class ShareDialog : public QDialog
 {
     Q_OBJECT
 
 public:
-    explicit ShareDialog(AccountPtr account, const QString &sharePath, const QString &localPath,
-                         bool resharingAllowed, QWidget *parent = 0);
+    explicit ShareDialog(AccountPtr account,
+                         const QString &sharePath,
+                         const QString &localPath,
+                         bool resharingAllowed,
+                         QWidget *parent = 0);
     ~ShareDialog();
+
     void getShares();
 
 private slots:
-    void slotSharesFetched(const QList<QSharedPointer<Share>> &shares);
-    void slotCreateShareFetched(const QSharedPointer<LinkShare> share);
-    void slotCreateShareRequiresPassword();
-    void slotDeleteShareFetched();
-    void slotPasswordSet();
-    void slotExpireSet();
-    void slotCalendarClicked(const QDate &date);
-    void slotCheckBoxShareLinkClicked();
-    void slotCheckBoxPasswordClicked();
-    void slotCheckBoxExpireClicked();
-    void slotPasswordReturnPressed();
-    void slotPasswordChanged(const QString& newText);
-    void slotPushButtonCopyLinkPressed();
-    void slotThumbnailFetched(const int &statusCode, const QByteArray &reply);
-    void slotCheckBoxEditingClicked();
-    void slotPublicUploadSet();
-
-    void displayError(int code, const QString &message);
-
     void done( int r );
-private:
-    void setShareCheckBoxTitle(bool haveShares);
-    void displayError(int code);
-    void displayError(const QString& errMsg);
-    void setShareLink( const QString& url );
-    void resizeEvent(QResizeEvent *e);
-    void redrawElidedUrl();
-    void setPublicUpload(bool publicUpload);
+    void slotThumbnailFetched(const int &statusCode, const QByteArray &reply);
 
+private:
     Ui::ShareDialog *_ui;
     AccountPtr _account;
     QString _sharePath;
     QString _localPath;
-    QString _shareUrl;
-#if 0
-    QString _folderAlias;
-    int     _uploadFails;
-    QString _expectedSyncFile;
-#endif
-
-    bool _passwordJobRunning;
-    void setPassword(const QString &password);
-    void setExpireDate(const QDate &date);
-
-    QProgressIndicator *_pi_link;
-    QProgressIndicator *_pi_password;
-    QProgressIndicator *_pi_date;
-    QProgressIndicator *_pi_editing;
-
-    ShareManager *_manager;
-    QSharedPointer<LinkShare> _share;
 
     bool _resharingAllowed;
-    bool _isFile;
+
+    ShareLinkWidget *_linkWidget;
+    ShareUserGroupWidget *_userGroupWidget;
 };
 
 }
