+++ src/gui/socketapi.h
@@ -75,7 +75,6 @@ private:
     Q_INVOKABLE void command_RETRIEVE_FOLDER_STATUS(const QString& argument, QIODevice* socket);
     Q_INVOKABLE void command_RETRIEVE_FILE_STATUS(const QString& argument, QIODevice* socket);
     Q_INVOKABLE void command_SHARE(const QString& localFile, QIODevice* socket);
-    Q_INVOKABLE void command_SHARE_USER_GROUP(const QString& localFile, QIODevice* socket);
 
     Q_INVOKABLE void command_VERSION(const QString& argument, QIODevice* socket);
